Extract common task code to :task:list-api
