+++ core/src/main/java/com/akjaw/core/common/domain/model/TimestampMilliseconds.kt
@@ -1,5 +1,6 @@
 package com.akjaw.core.common.domain.model
 
+//TODO maybe :stopwatch-api?
 data class TimestampMilliseconds(val value: Long) {
 
     infix operator fun plus(other: TimestampMilliseconds) =
