+++ features/stopwatch/build.gradle
@@ -43,6 +43,7 @@ android {
 
 dependencies {
     implementation project(":core")
+    implementation project(":features:task:list_api")
 
     implementation 'androidx.core:core-ktx:1.3.2'
     implementation 'androidx.appcompat:appcompat:1.2.0'
