+++ features/stopwatch/src/main/java/com/akjaw/stopwatch/view/AddStopwatchDialog.kt
@@ -12,12 +12,12 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import com.akjaw.core.common.domain.model.Task
 import com.akjaw.core.common.view.PaddedDialog
 import com.akjaw.core.common.view.theme.TimiComposeTheme
 import com.akjaw.core.common.view.theme.taskShape
 import com.akjaw.core.common.view.theme.taskTextColorFor
-import com.akjaw.core.common.view.theme.tasksPreview
+import com.akjaw.task.list_api.domain.Task
+import com.akjaw.task.list_api.view.tasksPreview
 
 @Composable
 internal fun AddStopwatchDialog(
