+++ features/stopwatch/src/test/java/com/akjaw/stopwatch/domain/StopwatchListOrchestratorTest.kt
@@ -1,9 +1,9 @@
 package com.akjaw.stopwatch.domain
 
 import androidx.compose.ui.graphics.Color
-import com.akjaw.core.common.domain.model.Task
 import com.akjaw.core.common.domain.model.toTimestampMilliseconds
 import com.akjaw.stopwatch.domain.model.StopwatchState
+import com.akjaw.task.list_api.Task
 import io.mockk.every
 import io.mockk.mockk
 import kotlinx.coroutines.CoroutineScope
@@ -26,8 +26,9 @@ import strikt.assertions.isNull
 internal class StopwatchListOrchestratorTest {
 
     companion object {
-        private val TASK1 = Task("First task", Color.White, false)
-        private val TASK2 = Task("Second the cooler task", Color.White, false)
+        private val TASK1 = com.akjaw.task.list_api.Task("First task", Color.White, false)
+        private val TASK2 =
+            com.akjaw.task.list_api.Task("Second the cooler task", Color.White, false)
     }
 
     private val stopwatchStateHolder: StopwatchStateHolder = mockk(relaxed = true) {
