+++ features/task/list/src/main/java/com/akjaw/task/list/composition/TaskListModule.kt
@@ -1,0 +1,16 @@
+package com.akjaw.task.list.composition
+
+import com.akjaw.task.list.data.TaskRepositoryImpl
+import com.akjaw.task.list_api.data.TaskRepository
+import dagger.Binds
+import dagger.Module
+import dagger.hilt.InstallIn
+import dagger.hilt.components.SingletonComponent
+
+@Module
+@InstallIn(SingletonComponent::class)
+internal abstract class TaskListModule {
+
+    @Binds
+    abstract fun bindTaskRepository(taskRepositoryImpl: TaskRepositoryImpl): TaskRepository
+}
