+++ settings.gradle
@@ -13,3 +13,4 @@ include ':features:stopwatch'
 include ':core'
 include ':features:task'
 include ':features:task:list'
+include ':features:task:list_api'
